Address review feedback: add role to thinking entries, extract shared parser, revert unintended unescapeWithMathProtection
- Add role: 'assistant' to thinking entries in normalizePersistedGeminiAssistantEntries
- Extract splitLegacyGeminiThoughtContent into shared/geminiThoughtParser.js to eliminate server/client duplication
- Revert unescapeWithMathProtection from handleSimpleAssistantMessage and structured-data text handler (was unintentionally added to previously untouched code paths)
