+++ shared/geminiThoughtParser.d.ts
@@ -1,0 +1,8 @@
+export type LegacyGeminiThoughtSegment = {
+  content: string;
+  isThinking: boolean;
+};
+
+export function splitLegacyGeminiThoughtContent(
+  text: string,
+): LegacyGeminiThoughtSegment[] | null;
